SetQuota to fix idempotency flaw
Existence of volume is not sufficient condition to succeed
provisioning as there may be not sufficient quota available
in the host. Therefore, succeed provisioning only iff quota
can be set on the (newly created/existing) volume.

(volume and PV are interchangeable in the below flow)

This fixes flaw in the following current flow:
1. Create volume 'a'
2. Set quota -> fails
3. Delete volume -> fails
4. Retry from k8s for creation of volume 'a'
5. Return success as volume 'a' exists

With the fix, step 5 would not succeed anymore, as Quobyte CSI driver
tries to set the Quota. With the fix retires are more seamless,
if customer increases sufficient tenant Quota, a future retry would
succeed.

+++ src/driver/controller.go
@@ -180,14 +180,22 @@ func (d *QuobyteDriver) CreateVolume(ctx context.Context, req *csi.CreateVolumeR
 		}
 	} else {
 		volUUID = volCreateResp.VolumeUuid
-		if createQuota {
-			err := quobyteClient.SetVolumeQuota(volUUID, capacity)
-			if err != nil {
-				// Volume is just created and volume database is empty. Therefore, use DeleteVolume
-				// call to delete the volume database and volume immediately (no need to erase any
-				// file data - so avoid erase API call)
-				quobyteClient.DeleteVolumeByResolvingNamesToUUID(volUUID, "")
-			}
+	}
+
+	// Creating a new volume/existence of volume alone is not sufficient when Quobyte
+	// tenant is configured with "disable_oversubscription: true"
+	// Creation of a quota ensures that provisioning succeeds only if there is sufficient Quota
+	// available for the request (must enable StorageClass createQuota).
+	// For shared volume, do not set Quota, as the requested Quota would be for directory
+	// (subpath of shared volume) - quota should be set by admin at tenant level for shared
+	// volumes
+	if !isSharedVolume && createQuota {
+		err := quobyteClient.SetVolumeQuota(volUUID, capacity)
+		if err != nil {
+			// Volume is just created and volume database is empty. Therefore, use DeleteVolume
+			// call to delete the volume database and volume immediately (no need to erase any
+			// file data - so avoid erase API call)
+			quobyteClient.DeleteVolumeByResolvingNamesToUUID(volUUID, "")
 			return nil, err
 		}
 	}
